Move ToastManager to dedicated top level package
