+++ admin/tabs/configs/differ/ConfigDifferModel.js
@@ -12,7 +12,8 @@ import {pluralize} from 'hoist/utils/JsUtils';
 import {XH, HoistModel} from 'hoist/core';
 import {LocalStore} from 'hoist/data';
 import {GridModel} from 'hoist/cmp/grid';
-import {MessageModel, ToastManager, StoreContextMenu} from 'hoist/cmp';
+import {MessageModel, StoreContextMenu} from 'hoist/cmp';
+import {ToastManager} from 'hoist/toast';
 import {baseCol} from 'hoist/columns/Core';
 import {nameCol} from 'hoist/admin/columns/Columns';
 import {p} from 'hoist/cmp/layout';
