+++ admin/tabs/services/ServiceModel.js
@@ -5,7 +5,7 @@
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
 import {XH, HoistModel} from 'hoist/core';
-import {ToastManager} from 'hoist/cmp';
+import {ToastManager} from 'hoist/toast';
 import {UrlStore} from 'hoist/data';
 import {GridModel} from 'hoist/cmp/grid';
 import {baseCol} from 'hoist/columns/Core';
