+++ toast/ToastManager.js
